requirements.txt
beautifulsoup4==4.11.1
langdetect==1.0.9
matplotlib==3.6.2
pandarallel==1.6.3
wordcloud==1.8.2.2
scikit-learn==1.1.3
pandas==1.5.1
pandas_gbq==0.15.0
protobuf==4.21.9
requests==2.28.1
selenium==4.6.0
streamlit==1.15.2
tqdm==4.64.1
webdriver_manager==3.8.4
openai==0.25.0
wandb==0.13.5
google-auth==2.14.1
google-cloud==0.34.0
